Enrichment analysis

Enrichment analysis

enrichment(
  hypeR_GEM_obj,
  genesets,
  genesets_name = "unknown",
  method = c("unweighted", "weighted"),
  weighted_by = "one_minus_fdr",
  a = 1,
  b = -1,
  sigmoid_transformation = TRUE,
  min_metabolite = 0,
  background = 1234567
)

Arguments

hypeR_GEM_obj

an hypeR_GEM object, the output of "signature2gene" function

genesets

a list of genesets

genesets_name

name of the geneset,e.g "KEGG"

method

enrichment method

weighted_by

the column name in the gene_table of hypeR_GEM_obj that represents the weight of each gene

a

proportional to the smoothness of the sigmoid function, default a = 1

b

the half-point threshold of the sigmoid function, default b = -1 (p-value = 0.1 as half-point)

sigmoid_transformation

logical; when method == "weighted", if TRUE apply .sigmoid_transformation() to `weighted_by`; if FALSE use the raw values

min_metabolite

minimum number of metabolite that drives the enrichment of a pathway

background

background parameter of hypergeometric test

Value

a list of type of statistical test and data