Map metabolic signatures to enzyme-coding genes

Map metabolic signatures to enzyme-coding genes

signature2gene(
  signatures,
  species = c("Human", "Mouse", "Rat", "Zebrafish", "Worm", "Other"),
  directional = FALSE,
  tables = NULL,
  merge = TRUE,
  reference_key = "refmet_name",
  promiscuous_threshold = NULL,
  ensemble_id = TRUE,
  background = NULL
)

Arguments

signatures

a list of metabolic signatures, each element is a data frame which has to contain a column with the same name as "reference_key"

species

the species of GEM model, if species == 'Other', a user-define GEM_table must be provided

directional

logical parameter, if TRUE, map metabolites to reactions where these metabolites are product only

tables

species == 'Other', a user-define GEM_table must be provided

merge

merged metabolites in different compartment, for example, MAM001_c, MAM001_e, ... -> MAM001

reference_key

the column name in the reference table which represents the standardized names (e.g. "refmet_name")

promiscuous_threshold

maximum allowable number of associated metabolites per gene; genes exceeding this threshold are excluded.

ensemble_id

if the genes in GEM_tables$gene_df is given by Ensemble, then ensemble_id = TRUE, otherwise, ensemble_id = FALSE

background

the background of the gene-specific hypergeometric test, default = NULL = number of merged metabolites

Value

a list containing two element: "mapped_metabolite_signatures" and "gene_tables" for each signature